+++ previous-versions/accounting/pkg/models/shared/billpayment.go
@@ -195,8 +195,8 @@ type BillPayment struct {
 	Metadata     *Metadata         `json:"metadata,omitempty"`
 	ModifiedDate *string           `json:"modifiedDate,omitempty"`
 	// Additional information associated with the payment.
-	Note             *string     `json:"note,omitempty"`
-	PaymentMethodRef interface{} `json:"paymentMethodRef,omitempty"`
+	Note             *string           `json:"note,omitempty"`
+	PaymentMethodRef *PaymentMethodRef `json:"paymentMethodRef,omitempty"`
 	// Additional information associated with the payment.
 	Reference          *string `json:"reference,omitempty"`
 	SourceModifiedDate *string `json:"sourceModifiedDate,omitempty"`
@@ -283,7 +283,7 @@ func (o *BillPayment) GetNote() *string {
 	return o.Note
 }
 
-func (o *BillPayment) GetPaymentMethodRef() interface{} {
+func (o *BillPayment) GetPaymentMethodRef() *PaymentMethodRef {
 	if o == nil {
 		return nil
 	}