+++ previous-versions/accounting/pkg/models/shared/paymentmethod.go
@@ -2,6 +2,42 @@
 
 package shared
 
+import (
+	"encoding/json"
+	"fmt"
+)
+
+// PaymentMethodStatus - Status of the Payment Method.
+type PaymentMethodStatus string
+
+const (
+	PaymentMethodStatusUnknown  PaymentMethodStatus = "Unknown"
+	PaymentMethodStatusActive   PaymentMethodStatus = "Active"
+	PaymentMethodStatusArchived PaymentMethodStatus = "Archived"
+)
+
+func (e PaymentMethodStatus) ToPointer() *PaymentMethodStatus {
+	return &e
+}
+
+func (e *PaymentMethodStatus) UnmarshalJSON(data []byte) error {
+	var v string
+	if err := json.Unmarshal(data, &v); err != nil {
+		return err
+	}
+	switch v {
+	case "Unknown":
+		fallthrough
+	case "Active":
+		fallthrough
+	case "Archived":
+		*e = PaymentMethodStatus(v)
+		return nil
+	default:
+		return fmt.Errorf("invalid value for PaymentMethodStatus: %v", v)
+	}
+}
+
 // PaymentMethod - > View the coverage for payment methods in the <a className="external" href="https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=paymentMethods" target="_blank">Data coverage explorer</a>.
 //
 // ## Overview